Refactor 2 + final milestones
Added an /obfuscation folder that holds all abstract classes. Each script is now an inheritor of an abstract class found here.
Added flask api support. This is accessible from root/app.py or root/app/flask_api_routes.py.
- The API will take user input and send the text to the frontend to generate the etymology. The API reads an HTML template and uses that as the webpage.
All scripts have been rewritten to be independent of one another, only making references to a class as an import.
Finalized a draft of the web scraper.
- It dynamically generates links to Wiktionary given the user's input. The web pages scraped sometimes return bad data, this is a known bug that requires more research.
- The webscraper will attempt to generate correct links, and has a failsafe built in to drop after third attempt at link generation.
The SQL script now properly writes to the database. Each time an input is taken in, it will add the new etymology to the database, decreasing load times each time a word is looked up.
Added a regex script for the web scraper to use.
Added Flask API documentation in the /root folder.
Added an HTML templates folder, which contains the HTML template for the Flask API to read

In the tests folder, there are tests for the following:
- clear_tables: drops the table in the SQL database. !!This will remove any learned data!!
- frontend_test: tests the parsing system
- retest: tests the regular expression used in the SQL script to make sure that no errant punctuation ends up in the etymology
- sql_test: a few lines of testing the SQL database to make sure it is reading/writing properly.
- testdb.sql: this is an original copy of the seeder database
- Flask API.postman_collection.json: a collection of Postman tests to check the Flask API. Can be run as a collection, or independently to see output.